+++ src/DBQuizActivityContollers/DBQuizTake.java
@@ -7,18 +7,18 @@ import java.sql.SQLException;
 import database.DBconnector;
 
 public class DBQuizTake {
-	
+
 	Connection connection;
-	
-	public DBQuizTake(){
-		
+
+	public DBQuizTake() {
+
 	}
-	
-	private void connect(){
+
+	private void connect() {
 		connection = new DBconnector().getConnection();
 	}
-	
-	private void disconnect(){
+
+	private void disconnect() {
 		try {
 			connection.close();
 		} catch (SQLException e) {
@@ -26,35 +26,28 @@ public class DBQuizTake {
 			e.printStackTrace();
 		}
 	}
-	
-	public void addActivity(int user_id, int quiz_id, int time_taken, int score){		
+
+	public void addActivity(int user_id, int quiz_id, int time_taken, double score) {
 		connect();
-		
-		String query = "insert into Quiz_taken(user_id, quiz_id, time_finished, time_taken, score) "
-				+ "values ("
-				+ user_id
-				+ ", "
-				+ quiz_id
-				+ ", current_timestamp() + interval '3' HOUR, "
-				+ time_taken
-				+ ", "
-				+ score
+
+		String query = "insert into Quiz_taken(user_id, quiz_id, time_finished, time_taken, score) " + "values ("
+				+ user_id + ", " + quiz_id + ", current_timestamp() + interval '3' HOUR, " + time_taken + ", " + score
 				+ ");";
 		PreparedStatement stm;
-		
+
 		System.out.println(query);
-		
-		try{
+
+		try {
 			stm = connection.prepareStatement(query);
 			stm.executeUpdate();
-					
+
 			connection.close();
-			
+
 		} catch (SQLException e) {
 			System.out.println("Error occured durin database connection!");
 			e.printStackTrace();
 		}
-		
+
 		disconnect();
 	}
 }
